requirements.txt
torch
numpy
Pillow
diffusers>=0.2.4
librosa
datasets
gradio
streamlit
